Move tests into cdm-image. Make jenkins cdm-test tests all pass.
